+++ R/misc.R
@@ -106,7 +106,7 @@ as_one_character <- function(x, allow_na = FALSE) {
   x <- as.character(x)
   if (length(x) != 1L || anyNA(x) && !allow_na) {
     s <- deparse_combine(s, max_char = 100L)
-    stop2("Cannot coerce ", s, " to a single character value.")
+    stop("Cannot coerce ", s, " to a single character value.")
   }
   x
 }